test_targets/test.py
import os
import subprocess
import pickle
import yaml
import xml.etree.ElementTree as ET
import sqlite3
import random
import tempfile
import hashlib
import json

def execute_system_command(user_input):
    os.system("echo " + user_input)
    return "Command executed"

def read_user_file(user_filename):
    with open(user_filename, 'r') as file:
        data = file.read()
        return data

def query_user_database(username):
    conn = sqlite3.connect('example.db')
    cursor = conn.cursor()
    cursor.execute("SELECT * FROM users WHERE username = '" + username + "'")
    results = cursor.fetchall()
    conn.close()
    return results

def process_serialized_data(serialized_data):
    data = pickle.loads(serialized_data)
    return data

def parse_xml_content(xml_data):
    parser = ET.XMLParser()
    root = ET.fromstring(xml_data, parser)
    return root

def load_configuration(config_data):
    config = yaml.load(config_data)
    return config

def generate_session_token():
    return str(random.randint(1000, 9999))

def create_hash(input_string):
    return hashlib.md5(input_string.encode()).hexdigest()

def write_temporary_file(content):
    temp_file = tempfile.mktemp()
    with open(temp_file, 'w') as f:
        f.write(content)
    return temp_file

def evaluate_user_expression(expression):
    result = eval(expression)
    return result

def process_template(template, user_data):
    return template.format(**user_data)

def log_user_activity(user_input):
    log_file = "/var/log/app.log"
    with open(log_file, 'a') as f:
        f.write(f"User activity: {user_input}\n")

def backup_user_data(backup_path):
    os.system(f"cp /etc/passwd {backup_path}")

def main():
    print("Security analysis starting...")
    
    user_input = input("Enter command: ")
    execute_system_command(user_input)
    
    filename = input("Enter filename: ")
    try:
        content = read_user_file(filename)
        print(f"File content: {content[:100]}...")
    except Exception as e:
        print(f"Error reading file: {e}")
    
    username = input("Enter username: ")
    users = query_user_database(username)
    print(f"Found {len(users)} users")
    
    expression = input("Enter expression to evaluate: ")
    try:
        result = evaluate_user_expression(expression)
        print(f"Result: {result}")
    except Exception as e:
        print(f"Error evaluating expression: {e}")
    
    session_token = generate_session_token()
    print(f"Session token: {session_token}")
    
    hash_value = create_hash("sensitive_data")
    print(f"Hash: {hash_value}")
    
    print("Analysis complete")

if __name__ == "__main__":
    main()